repositories: dedupe logic
This change unifies shared logic across repository implementations and
improves inheritance model.
