+++ src/poetry/console/commands/show.py
@@ -14,8 +14,8 @@ if TYPE_CHECKING:
     from poetry.core.packages.package import Package
 
     from poetry.packages.project_package import ProjectPackage
-    from poetry.repositories import Repository
     from poetry.repositories.installed_repository import InstalledRepository
+    from poetry.repositories.repository import Repository
 
 
 class ShowCommand(GroupCommand):
