--- src/poetry/repositories/remote_repository.py
@@ -1,18 +1,0 @@
-from __future__ import annotations
-
-from poetry.repositories.repository import Repository
-
-
-class RemoteRepository(Repository):
-    def __init__(self, url: str) -> None:
-        self._url = url
-
-        super().__init__()
-
-    @property
-    def url(self) -> str:
-        return self._url
-
-    @property
-    def authenticated_url(self) -> str:
-        return self._url
